% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVR.r
\name{calc.mvd}
\alias{calc.mvd}
\title{Calculate multivariate richness and functional overlap}
\usage{
calc.mvd(in.mat, in.com, traitspace, calc.ovr)
}
\arguments{
\item{in.mat}{A record by trait matrix}

\item{in.com}{A record by community matrix}

\item{traitspace}{The total trait space for scaling purposes}

\item{calc.ovr}{An indicator for whether overlap metrics should be calculated}
}
\description{
Function to reduce to functional units, and calculate multivariate richness
and functional overlap
}
