% Generated by roxygen2 (4.0.1): do not edit by hand
\name{check.breaks}
\alias{check.breaks}
\title{Check breaks}
\usage{
check.breaks(breaks, in.mat)
}
\arguments{
\item{breaks}{A list containing break points to use for each continuous trait.
Categorical traits are unmodified, and should be a list containing the text "cat"}

\item{in.mat}{A record x trait matrix}
}
\description{
Check that break point lists are all the same length
}

