% Generated by roxygen2 (4.0.1): do not edit by hand
\name{check.traits}
\alias{check.traits}
\title{Check trait}
\usage{
check.traits(in.mat, tr1.lim, tr2.lim, sim.type, filter.vals = "none")
}
\arguments{
\item{in.mat}{A record x trait matrix}

\item{tr1}{Trait values for trait 1}

\item{tr2}{Trait values for trait 2}

\item{tr1.lim}{The minimum and maximum allowable values for trait 1}

\item{tr2.lim}{The minimum and maximum allowable values for trait 2}

\item{sim.type}{"random" runs random assembly (or if any sim.type other than limiting or filter is specified), "limiting" runs limiting similarity, and "filter" applies an environmental filter.}

\item{filter.vals}{Vector of filter values in text form with a semi-colon delimiter (sorry!) e.g., c("1,1", "1;2") would only allow values of 1,1 and 2,2 in the data.}
}
\description{
Check trait for compatibility with assembly procedure & adjust as necessary
}
\note{
This function is very inefficient & could be optimized!
}

