% Generated by roxygen2 (4.0.1): do not edit by hand
\name{expand.breaks}
\alias{expand.breaks}
\title{Expand trait breaks}
\usage{
expand.breaks(in.lst, target.reps)
}
\arguments{
\item{in.lst}{the list of trait breakpoints to be expanded}

\item{target.reps}{the target length of the trait breakpoint list.}
}
\value{
an expanded list of trait breakpoints
}
\description{
Function to match longer lists of trait breaks. This may be a sub-optimal solution
}

