% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVR.r
\name{get.breaks}
\alias{get.breaks}
\title{Function to calculate break points for a trait}
\usage{
get.breaks(min.val, max.val)
}
\arguments{
\item{min.val}{The minimum value to be used. Can be taken from the data, or can be based on a priori knowledge}

\item{max.val}{The maximum value to be used in generating break points. Can be taken from the data or can be based on a priori knowledge.}
}
\description{
takes range of values, and then creates break points for all possible sub-categorizations of those values
WARNING: Currently only gives breakpoints for integer values!
}
