% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.traitspace}
\alias{get.traitspace}
\title{Calculate trait space (simple)}
\usage{
get.traitspace(in.mat, col.res, col.mins, col.maxs)
}
\arguments{
\item{in.mat}{A record x trait matrix}

\item{col.res}{The number of decimal places to use for rounding (resolution) purposes}

\item{col.mins}{The minimum values to use for the trait space. "use data" extracts minimums from in.mat}

\item{col.maxs}{The maximum values to use for the trait space. "use data" extracts maximums from in.mat}
}
\description{
Function to calculate the traitspace using minimum and maximum column values.
Creates a hypercubic traitspace.
}

