% Generated by roxygen2 (4.0.1): do not edit by hand
\name{listtotext}
\alias{listtotext}
\title{List to text}
\usage{
listtotext(inlist, separator)
}
\arguments{
\item{inlist}{An unnested list}

\item{separator}{A separator to separate elements of the list.}
}
\description{
Take a list (or vector) and convert it to text separated by separator.
Does not work for nested lists
ORIGINALLY FROM MYR.R script
}

