% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make.trait}
\alias{make.trait}
\title{Make trait}
\usage{
make.trait(trlim, n.recs)
}
\arguments{
\item{trlim}{A vector containing the minimum and maximum values for the trait}

\item{n.recs}{The number of trait values to create for trait}
}
\description{
Create integer trait values from a uniform random distribution
}

