% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVR.r
\name{matrix.check}
\alias{matrix.check}
\title{Check that input to mvfd is in matrix format}
\usage{
matrix.check(in.mat)
}
\arguments{
\item{in.mat}{An input to be tested for matrix formatting}
}
\description{
Check that input to mvfd is in matrix format
}
