% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVR.r
\docType{package}
\name{multirich-package}
\alias{multirich-package}
\title{multirich (package)}
\description{
Package to calculate the number of unique trait combinations and scale the
number of unique trait combinations by the total number of trait combinations
possible as measures of multivariate richness.
}
\details{
\tabular{ll}{ %%Note tabular{ll} is tabular{lowercase(LL)} not tabular{11}
Package: \tab multirich\cr
Type: \tab Package\cr
Version: \tab 2.1.1\cr
Date: \tab 2015-09-14\cr
License: \tab GPL-2 (or later)\cr
}
}
\references{
Keyel, A.C. and K. Wiegand. (in review) Validating the use of
Unique Trait Combinations for measuring multivariate functional richness.
}
\author{
Alexander "Sasha" Keyel\cr
Maintainer: Sasha Keyel <skeyel@gmail.com>
}
\keyword{package}
