% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVR.r
\name{plot.sensitivity}
\alias{plot.sensitivity}
\title{Plot sensitivity analysis results}
\usage{
\method{plot}{sensitivity}(sutc.vals, traitspaces, n, com.names = "none", in.pdf = "none")
}
\arguments{
\item{sutc.vals}{A list containing scaled unique trait combinations (see above)}

\item{traitspaces}{A vector containing the traitspace for each combination value}

\item{n}{The sample size of the trait x record matrix}

\item{com.names}{The names of the communities being evaluated (for labeling purposes)}

\item{in.pdf}{A pdf file to be generated, containing the two plots. "none" skips generating a pdf, and "no.plot" disables the function}
}
\description{
Goal is to create a plot similar to a ROC plot, where at one side, everything is in a single
category, and on the other end, everything is in its own category.
}
