% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVR.r
\name{sensitivity.boxplots}
\alias{sensitivity.boxplots}
\title{Make boxplots based on multiple simulations}
\usage{
sensitivity.boxplots(results.mat, traitspaces, n.recs)
}
\arguments{
\item{results.mat}{An input matrix with ncol equal to the number of elements in traitspaces,
and a given row containing the values for each trait space for that simulation run.}

\item{traitspaces}{A vector identifying the traitspaces at which the simulation analysis was carried out}

\item{n.recs}{The number of records in the original matrix.
This is the point where if every record were unique, the entire traitspace would be full.}
}
\description{
Takes the output results from multiple simulation iterations and plots boxplots displaying the results
}
