% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVR.r
\name{utc.sim}
\alias{utc.sim}
\title{Sensitivity Simulator}
\usage{
utc.sim(
  n.reps,
  n.recs,
  tr1.lim,
  tr2.lim,
  sim.type = "random",
  filter.vals = "none"
)
}
\arguments{
\item{n.reps}{The number of replicates for the simulation}

\item{n.recs}{The number of records in the record x trait matrix}

\item{tr1.lim}{The minimum and maximum allowable values for trait 1}

\item{tr2.lim}{The minimum and maximum allowable values for trait 2}

\item{sim.type}{"random" runs random assembly (or if any sim.type other than limiting or filter is specified), "limiting" runs limiting similarity, and "filter" applies an environmental filter.}

\item{filter.vals}{Vector of filter values in text form with a semi-colon delimiter (sorry!) e.g., c("1,1", "1;2") would only allow values of 1,1 and 2,2 in the data.}
}
\description{
Function to help run simulations for sensitivity analysis.
Note that for simplicity, the number of traits is restricted to two, each with the same range of values, as can be done for real traits by standardizing by range.
}
