% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calc.mvo}
\alias{calc.mvo}
\title{Calculate functional overlap}
\usage{
calc.mvo(in.mat, dups)
}
\arguments{
\item{in.mat}{a record x trait matrix}

\item{dups}{dups indicate which records in the record x trait matrix are duplicates}
}
\description{
Calculate functional overlap
}

