% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVR.r
\name{df.categorize}
\alias{df.categorize}
\title{Categorize data}
\usage{
df.categorize(in.mat, cell.res)
}
\arguments{
\item{in.mat}{A record x trait matrix}

\item{cell.res}{The number of decimal places to round the data to.}
}
\description{
Currently by rounding, which is a sub-optimal approach for many questions
}
