% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{multirich-package}
\alias{multirich-package}
\title{multirich (package)}
\description{
Package to calculate the number of unique trait combinations and scale the
number of unique trait combinations by the total number of trait combinations
possible as measures of multivariate richness.
}
\details{
\tabular{ll}{ %%Note tabular{ll} is tabular{lowercase(LL)} not tabular{11}
Package: \tab multirich\cr
Type: \tab Package\cr
Version: \tab 2.0.2\cr
Date: \tab 2014-02-04\cr
License: \tab GPL-2 (or later)\cr
}
}
\author{
Alexander "Sasha" Keyel\cr
Maintainer: Sasha Keyel <skeyel@gmail.com>
}
\references{
Keyel, A.C. and K. Wiegand. (in review) Validating the use of
Unique Trait Combinations for measuring multivariate functional richness.
}
\keyword{package}

